requirements.txt
colorama
websocket-client
markupsafe==2.0.1
ndg-httpsclient
pyopenssl
pyasn1
pycryptodome
datetime
colorama
requests